+++ testing/rspec/spec/integration/adapters/gherkin_5_adapter_spec.rb
@@ -1,9 +1,9 @@
 require "#{File.dirname(__FILE__)}/../../spec_helper"
 
 
-describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
+describe 'Gherkin4Adapter, Integration', :if => gherkin?(5) do
 
-  let(:clazz) { CukeModeler::Gherkin3Adapter }
+  let(:clazz) { CukeModeler::Gherkin5Adapter }
   let(:adapter) { clazz.new }
   let(:source_text) { "# feature comment
                        @tag1 @tag2 @tag3
@@ -68,23 +68,22 @@ describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
                            | value |
                        # final comment" }
   let(:feature_file_model) { test_file_path = CukeModeler::FileHelper.create_feature_file(:text => source_text, :name => 'adapter_test_file')
-                             CukeModeler::FeatureFile.new(test_file_path) }
+  CukeModeler::FeatureFile.new(test_file_path) }
   let(:feature_model) { feature_file_model.feature }
 
 
   it "does not store parsing data for a feature file's children" do
     model = feature_file_model
 
-    expect(model.parsing_data).to be_nil
+    expect(model.parsing_data[:comments]).to be_nil
+    expect(model.parsing_data[:feature]).to be_nil
   end
 
   it "does not store parsing data for a feature's children" do
     model = feature_model
 
-    expect(model.parsing_data[:comments]).to be_nil
     expect(model.parsing_data[:tags]).to be_nil
-    expect(model.parsing_data[:scenarioDefinitions]).to be_nil
-    expect(model.parsing_data[:background]).to be_nil
+    expect(model.parsing_data[:children]).to be_nil
   end
 
   it "does not store parsing data for a background's children" do
@@ -150,13 +149,13 @@ describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
   describe 'stuff that is in no way part of the public API and entirely subject to change' do
 
     it 'provides a useful explosion message if it encounters an entirely new type of test' do
-      partial_feature_ast = {:type => :Feature, :location => {:line => 1, :column => 1}, :scenarioDefinitions => [{:type => :some_unknown_type}]}
+      partial_feature_ast = { :type => :Feature, :location => { :line => 1, :column => 1 }, :children => [{ :type => :some_unknown_type }] }
 
       expect { adapter.adapt_feature!(partial_feature_ast) }.to raise_error(ArgumentError, /Unknown.*some_unknown_type/)
     end
 
     it 'provides a useful explosion message if it encounters an entirely new type of step block' do
-      partial_feature_ast = {:type => :Feature, :location => {:line => 1, :column => 1}, :scenarioDefinitions => [{:type => :Scenario, :tags => [], :location => {:line => 1, :column => 1}, :steps => [{:type => :Step, :location => {:line => 1, :column => 1}, :argument => {:type => :some_unknown_type, :location => {:line => 1, :column => 1}, :content => ""}}]}]}
+      partial_feature_ast = { :type => :Feature, :location => { :line => 1, :column => 1 }, :children => [{ :type => :Scenario, :tags => [], :location => { :line => 1, :column => 1 }, :steps => [{ :type => :Step, :location => { :line => 1, :column => 1 }, :argument => { :type => :some_unknown_type, :location => { :line => 1, :column => 1 }, :content => "" } }] }] }
 
       expect { adapter.adapt_feature!(partial_feature_ast) }.to raise_error(ArgumentError, /Unknown.*some_unknown_type/)
     end