+++ testing/rspec/spec/integration/adapters/gherkin_9_adapter_spec.rb
@@ -1,9 +1,9 @@
 require "#{File.dirname(__FILE__)}/../../spec_helper"
 
 
-describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
+describe 'Gherkin9Adapter, Integration', :if => gherkin?(9) do
 
-  let(:clazz) { CukeModeler::Gherkin3Adapter }
+  let(:clazz) { CukeModeler::Gherkin9Adapter }
   let(:adapter) { clazz.new }
   let(:source_text) { "# feature comment
                        @tag1 @tag2 @tag3
@@ -67,53 +67,55 @@ describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
                            | param |
                            | value |
                        # final comment" }
-  let(:feature_file_model) { test_file_path = CukeModeler::FileHelper.create_feature_file(:text => source_text, :name => 'adapter_test_file')
-                             CukeModeler::FeatureFile.new(test_file_path) }
+  let(:feature_file_model) do
+    test_file_path = CukeModeler::FileHelper.create_feature_file(:text => source_text, :name => 'adapter_test_file')
+
+    CukeModeler::FeatureFile.new(test_file_path)
+  end
   let(:feature_model) { feature_file_model.feature }
 
 
   it "does not store parsing data for a feature file's children" do
     model = feature_file_model
 
-    expect(model.parsing_data).to be_nil
+    expect(model.parsing_data[:comments]).to be_nil
+    expect(model.parsing_data[:feature]).to be_nil
   end
 
   it "does not store parsing data for a feature's children" do
     model = feature_model
 
-    expect(model.parsing_data[:comments]).to be_nil
     expect(model.parsing_data[:tags]).to be_nil
-    expect(model.parsing_data[:scenarioDefinitions]).to be_nil
-    expect(model.parsing_data[:background]).to be_nil
+    expect(model.parsing_data[:children]).to be_nil
   end
 
   it "does not store parsing data for a background's children" do
     model = feature_model.background
 
-    expect(model.parsing_data[:steps]).to be_nil
+    expect(model.parsing_data[:background][:steps]).to be_nil
   end
 
   it "does not store parsing data for a scenario's children" do
     model = feature_model.scenarios.first
 
-    expect(model.parsing_data[:tags]).to be_nil
-    expect(model.parsing_data[:steps]).to be_nil
+    expect(model.parsing_data[:scenario][:tags]).to be_nil
+    expect(model.parsing_data[:scenario][:steps]).to be_nil
   end
 
   it "does not store parsing data for an outline's children" do
     model = feature_model.outlines.first
 
-    expect(model.parsing_data[:tags]).to be_nil
-    expect(model.parsing_data[:steps]).to be_nil
-    expect(model.parsing_data[:examples]).to be_nil
+    expect(model.parsing_data[:scenario][:tags]).to be_nil
+    expect(model.parsing_data[:scenario][:steps]).to be_nil
+    expect(model.parsing_data[:scenario][:examples]).to be_nil
   end
 
   it "does not store parsing data for an example's children" do
     model = feature_model.outlines.first.examples.first
 
     expect(model.parsing_data[:tags]).to be_nil
-    expect(model.parsing_data[:tableHeader]).to be_nil
-    expect(model.parsing_data[:tableBody]).to be_nil
+    expect(model.parsing_data[:table_header]).to be_nil
+    expect(model.parsing_data[:table_body]).to be_nil
   end
 
   it "does not store parsing data for an example row's children" do
@@ -125,13 +127,13 @@ describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
   it "does not store parsing data for a step's children, table" do
     model = feature_model.outlines.first.steps.first
 
-    expect(model.parsing_data[:argument]).to be_nil
+    expect(model.parsing_data[:data_table]).to be_nil
   end
 
   it "does not store parsing data for a step's children, doc string" do
     model = feature_model.outlines.first.steps.last
 
-    expect(model.parsing_data[:argument]).to be_nil
+    expect(model.parsing_data[:doc_string]).to be_nil
   end
 
   it "does not store parsing data for a table's children" do
@@ -150,13 +152,7 @@ describe 'Gherkin3Adapter, Integration', :if => gherkin?(3) do
   describe 'stuff that is in no way part of the public API and entirely subject to change' do
 
     it 'provides a useful explosion message if it encounters an entirely new type of test' do
-      partial_feature_ast = {:type => :Feature, :location => {:line => 1, :column => 1}, :scenarioDefinitions => [{:type => :some_unknown_type}]}
-
-      expect { adapter.adapt_feature!(partial_feature_ast) }.to raise_error(ArgumentError, /Unknown.*some_unknown_type/)
-    end
-
-    it 'provides a useful explosion message if it encounters an entirely new type of step block' do
-      partial_feature_ast = {:type => :Feature, :location => {:line => 1, :column => 1}, :scenarioDefinitions => [{:type => :Scenario, :tags => [], :location => {:line => 1, :column => 1}, :steps => [{:type => :Step, :location => {:line => 1, :column => 1}, :argument => {:type => :some_unknown_type, :location => {:line => 1, :column => 1}, :content => ""}}]}]}
+      partial_feature_ast = { :type => :Feature, :location => { :line => 1, :column => 1 }, :children => [{ :some_unknown_type => {} }] }
 
       expect { adapter.adapt_feature!(partial_feature_ast) }.to raise_error(ArgumentError, /Unknown.*some_unknown_type/)
     end