+++ testing/rspec/spec/integration/parsing_integration_spec.rb
@@ -21,15 +21,11 @@ describe 'Parsing, Integration' do
       end
     end
 
-    it 'loads the correct dialects based on the version of Gherkin used', :gherkin6 => true do
+    it 'loads the correct dialects based on the version of Gherkin used', :unless => gherkin?(2) do
       expect(nodule.dialects).to equal(Gherkin::DIALECTS)
     end
 
-    it 'loads the correct dialects based on the version of Gherkin used', :gherkin3 => true, :gherkin4_5 => true do
-      expect(nodule.dialects).to equal(Gherkin::DIALECTS)
-    end
-
-    it 'loads the correct dialects based on the version of Gherkin used', :gherkin2 => true do
+    it 'loads the correct dialects based on the version of Gherkin used', :if => gherkin?(2) do
       expect(nodule.dialects).to equal(Gherkin::I18n::LANGUAGES)
     end
 
@@ -112,6 +108,31 @@ describe 'Parsing, Integration' do
               define_method(:parsing_method, $old_method)
             end
           end
+        end
+      end
+
+    end
+
+    describe 'parsing invalid Gherkin' do
+
+      it 'correctly bubbles up parsing errors', :if => gherkin?(6, 7, 8, 9) do
+        expect { nodule.parse_text('bad file') }.to raise_error(/RuntimeError.*#EOF/)
+      end
+
+      it 'correctly bubbles up parsing errors', :if => gherkin?(4, 5) do
+        expect { nodule.parse_text('bad file') }.to raise_error(/Gherkin::CompositeParserException.*#EOF/m)
+      end
+
+      it 'correctly bubbles up parsing errors', :if => gherkin?(3) do
+        expect { nodule.parse_text('bad file') }.to raise_error(/Gherkin::CompositeParserException.*unexpected end of file/m)
+      end
+
+      it 'correctly bubbles up parsing errors', :if => gherkin?(2) do
+        # A different error is thrown on JRuby
+        if RUBY_PLATFORM == "java"
+          expect { nodule.parse_text('bad file') }.to raise_error(/Java::GherkinLexer.*_FEATURE_END_/m)
+        else
+          expect { nodule.parse_text('bad file') }.to raise_error(/Gherkin::Lexer::LexingError.*error on line 1/)
         end
       end