+++ CHANGELOG.md
@@ -6,8 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Nothing yet...
+### Added
+ - Support added for more versions of `gherkin`
+   - 9.x
+   - 8.x
+   - 7.x
 
+### Fixed
+ - Parsing errors are now correctly bubbled up when using Gherkin 6.x
 
 ## [2.0.0] - 2020-02-11